requirements.txt
Django>=2.0, <3.0
djangorestframework==3.11
djangorestframework-api-key==1.4.1  
psycopg2>=2.7, <3.0